# openapi.yaml
openapi: 3.1.0
info:
  title: Ekklesia Hydra Middleware API
  version: 0.3.0
  description: |
    Express API middleware that manages voting operations inside a Hydra L2
    state channel on Cardano. Handles ballot minting on L1, in-head voter
    registration/voting via TRP, evidence pinning to IPFS, and settlement
    (finalize → burn → close).

    All responses follow a uniform envelope:
      - Success: `{ "status": "SUCCESS", "data": { ... } }`
      - Error:   `{ "status": "ERROR", "code": "...", "message": "..." }`

    All endpoints require the `x-api-key` header except where otherwise noted
    (the middleware applies globally).
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://{host}
    description: Deployed instance
    variables:
      host:
        default: hydra.example.com
security:
  - ApiKeyAuth: []
...
paths:
...
  /head-info:
    get:
      tags: [Head Lifecycle]
      summary: Full HydraHeadInfo
      responses:
        '200':
          description: Full head info from the monitor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '503': { $ref: '#/components/responses/Unavailable' }
  /start:
    post:
      tags: [Head Lifecycle]
      summary: Commit UTxOs and open head
      description: |
        Commits the listed UTxOs (including the (601) ballot token) and waits
        for the head to reach Open status. Optionally fetches and caches the
        ballot definition from IPFS. Caches ballot identity (policy/token) for
        voting routes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [utxos]
              properties:
                utxos:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [txHash, outputIndex]
                    properties:
                      txHash: { type: string }
                      outputIndex: { type: integer, minimum: 0 }
                ballotIpfsCid:
                  type: string
                  description: Ballot definition CID returned by /prepare. Caches the ballot.
                ballotPolicy:
                  type: string
                  description: Hex policy ID of the ballot tokens.
                ballotToken:
                  type: string
                  description: Hex instance asset name of the (601) token.
                resultsAddress:
                  type: string
                  description: |
                    Destination for the (601) token after the finalize tx.
                    Cached server-side and passed into `finalize_ballot` as
                    `results_address`. If omitted, the finalize output falls
                    back to the admin payment address.
      responses:
        '200':
          description: Head opened
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          ballotCached: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /close:
    post:
      tags: [Head Lifecycle]
      summary: Close the head (requires close token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [closeToken]
              properties:
                closeToken: { type: string }
      responses:
        '200':
          description: Head closed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          closed: { type: boolean }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ---------- Queue ----------
...
  /queue/drain:
    post:
      tags: [Queue]
      summary: Block until queue is empty
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                timeoutMs: { type: integer, default: 600000 }
      responses:
        '200':
          description: Queue drained
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          drained: { type: boolean }
                          message: { type: string }

  # ---------- Voting ----------
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  responses:
    BadRequest:
      description: Missing or invalid fields
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    InternalError:
      description: Internal error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unavailable:
      description: Downstream dependency (Hydra node / IPFS / wallet) unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status: { type: string, enum: [SUCCESS] }
        data: { type: object }
    ErrorEnvelope:
      type: object
      required: [status, code, message]
      properties:
        status: { type: string, enum: [ERROR] }
        code:
          type: string
          description: Machine-readable error code.
          enum:
            - UNAUTHORIZED
            - MISSING_FIELDS
            - INVALID_INPUT
            - INVALID_VOTER_ID
            - SIGNATURE_INVALID
            - CONFLICT
            - NO_BALLOT_CACHED
            - WALLET_INIT_FAILED
            - CLIENT_INIT_FAILED
            - CLOSE_TOKEN_INVALID
            - IPFS_UNAVAILABLE
            - HYDRA_UNREACHABLE
            - INTERNAL_ERROR
        message: { type: string }

    # ---------- Ballot Definition ----------
...